fixed traffic finding bugs

+++ app/src/main/java/Modules/Route.java
@@ -12,6 +12,6 @@ public class Route {
     public LatLng endLocation;
     public String startAddress;
     public LatLng startLocation;
-    public int bingCongestionScore;
+    public double bingCongestionScore;
     public List<LatLng> points;
 }
